UI
- implemented recommendation UI element,
- added recommendations to catalog product page,
- fixed nav drawer bug,
- implemented disposal of unfinished  network requests on catalog product page,
- (refactoring) continued replacement of qt-specific numeric types with standard ones

+++ widgets/storelistitem.h
@@ -23,13 +23,13 @@ public:
     ~StoreListItem();
 
 signals:
-    void navigateToProduct(quint64 id);
+    void navigateToProduct(unsigned long long id);
 
 protected:
     void mousePressEvent(QMouseEvent *event) override;
 
 private:
-    quint64 itemId;
+    unsigned long long itemId;
     QNetworkReply *imageReply;
     Ui::StoreListItem *ui;
 };
